+++ src/Foundation/Models/Order.php
@@ -16,6 +16,8 @@ namespace Vanilo\Foundation\Models;
 
 use Illuminate\Database\Eloquent\Builder;
 use Illuminate\Database\Eloquent\Relations\BelongsTo;
+use Vanilo\Channel\Models\Channel;
+use Vanilo\Channel\Models\ChannelProxy;
 use Vanilo\Contracts\Payable;
 use Vanilo\Order\Models\Order as BaseOrder;
 use Vanilo\Payment\Contracts\Payment;
@@ -23,6 +25,8 @@ use Vanilo\Payment\Models\PaymentProxy;
 
 /**
  * @property null|Payment $currentPayment is only available if order was fetched with withCurrentPayment() scope
+ * @property null|int $channel_id
+ * @property null|Channel $channel
  */
 class Order extends BaseOrder implements Payable
 {
@@ -49,6 +53,11 @@ class Order extends BaseOrder implements Payable
     public function getCurrency(): string
     {
         return config('vanilo.foundation.currency.code');
+    }
+
+    public function channel(): BelongsTo
+    {
+        return $this->belongsTo(ChannelProxy::modelClass());
     }
 
     public function getCurrentPayment(): ?Payment