+++ src/Foundation/resources/database/migrations/2022_07_05_154855_add_channel_id_to_orders_table.php
@@ -1,0 +1,24 @@
+<?php
+
+declare(strict_types=1);
+
+use Illuminate\Database\Migrations\Migration;
+use Illuminate\Database\Schema\Blueprint;
+use Illuminate\Support\Facades\Schema;
+
+return new class () extends Migration {
+
+    public function up()
+    {
+        Schema::table('orders', function (Blueprint $table) {
+            $table->intOrBigIntBasedOnRelated('channel_id', Schema::connection(null), 'channels.id')->unsigned()->nullable();
+        });
+    }
+
+    public function down()
+    {
+        Schema::table('orders', function (Blueprint $table) {
+            $table->dropColumn('channel_id');
+        });
+    }
+};